pyproject.toml
[project]
name = "jira-script"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jira==3.8.0",
    "requests==2.28.1",
    "html2text",
    "PyPDF2",
    "reportlab",
    "google-api-python-client",
    "google-auth-httplib2",
    "google-auth-oauthlib",
    "python-dotenv==1.0.1",
    "pydantic==2.4.2",
    "pydantic-settings==2.0.2",
    "pillow==10.4.0",
    "atlassian-python-api==3.41.16",
  
]
